Fix SQL Complaint and finalize modularization
