+++ Communities/app/controllers/accountCreation.go
@@ -11,7 +11,7 @@ import (
 	"golang.org/x/crypto/bcrypt"
 )
 
-// Used as a way to control renders
+// Defining the controller used in this file
 type AccountCreation struct {
 	*revel.Controller
 }
