+++ Communities/app/controllers/profile.go
@@ -10,7 +10,7 @@ import (
 	_ "github.com/go-sql-driver/mysql"
 )
 
-// Used as a way to control renders
+// Defining the controller used in this file
 type Profile struct {
 	*revel.Controller
 }
