+++ Communities/app/views/Communities.html
@@ -1,0 +1,11 @@
+
+			<div class = "communityWindow">
+				<b>Modular Bois</b><br/>
+				Testing the modularity of this application
+				<form action="/LoadAssociatedData" method="POST" right="1%">
+					<input type="hidden" name="CurrentCommunity" value="Modular Bois" >
+					<input type="hidden" name="CurrentCommunityDescription" value="Testing the modularity of this application">
+					<button type="submit">Visit Community</button><br>
+				</form>
+		    </div>
+
