+++ Communities/app/views/Community/CreateCommunity.html
@@ -1,0 +1,83 @@
+<!--
+    Community Creation Page
+    This page is responsible for
+    displaying the form that is 
+    responsible for constructing communities
+-->
+
+<!--Setting title and including default templates-->
+{{set . "title" "New Community  - Communities"}}
+{{template "header.html" .}}
+
+<!--Template responsible for displaying error and success messages-->
+{{template "flash.html" .}}
+
+<!--
+    Defining setups for inputs, buttons
+    and displays
+-->
+<style>
+.map{
+    position: absolute;
+    width: 50%;
+    height: 100%;
+    right: 0%;
+  }
+  .newCommunityData{
+    padding-left: 10px;
+    width: 50%;
+    height: 100%;
+    border: 10px solid black;
+    left: 1%;
+  }
+input[type=text]{   
+        width: 50%;   
+        margin: 8px 0;  
+        padding: 12px 20px;   
+        display: inline-block;   
+        border: 2px solid rgb(0, 0, 0);   
+        box-sizing: border-box;   
+}
+button {   
+        background-color: #000000;   
+        width: 20%;  
+        color: rgb(184, 182, 179);   
+        padding: 15px;   
+        margin: 10px 0px;   
+        border: none;   
+        cursor: pointer;   
+    }  
+button:hover {   
+        opacity: 0.7;   
+    } 
+</style>
+
+<!--Display for the map-->
+<div class = "map">
+  <img src="/public/img/my-map.png" height = 100%> 
+</div>
+
+<!--Display for the form-->
+<div class = "newCommunityData">
+    <form action = "/ConstructCommunity" method="POST">
+        <label>
+            Select an place in the map where you want to place your community.
+        </label><br/>
+        <!--Prompting the user for what he/she wants to call the community-->
+        <label> What do you want to call your community?</label><br/>
+        <input type="text" name = "NewCommunityName" maxlength = "24" required><br/>
+        <!--Prompting the user for the description of the community-->
+        <label> Enter a description about your community.</label><br/>
+        <input type="text" name = "CommunityDescription" maxlength = "250" required><br/>
+        <!--Button to submit all inputted data for the form-->
+        <button type="submit">Create Community</button>
+    </form>
+
+    <!--Go back button that returns to home page-->
+    <form action = "/home" method = "GET">
+        <button type="submit">Go Back</button>
+    </form>
+</div>
+
+
+{{template "footer.html" .}}
