+++ Communities/app/views/Community/NewEvent.html
@@ -1,0 +1,68 @@
+<!--
+    This page is responsible for
+    handling the creation of new events
+-->
+{{set . "title" "New Event  - Communities"}}
+{{template "header.html" .}}
+
+<!--
+    Defining setups for inputs, buttons, and containers
+-->
+<style>
+    input[type=text]{   
+            width: 100%;   
+            margin: 8px 0;  
+            padding: 12px 20px;   
+            display: inline-block;   
+            border: 2px solid rgb(0, 0, 0);   
+            box-sizing: border-box;   
+    }
+    button {   
+            background-color: #000000;   
+            width: 25%;  
+            color: rgb(184, 182, 179);   
+            padding: 15px;   
+            margin: 10px 0px;   
+            border: none;   
+            cursor: pointer;   
+        }  
+    button:hover {   
+            opacity: 0.7;   
+        } 
+    .container {   
+            padding: 25px; 
+            border: 3px solid #f1f1f1;   
+    
+        } 
+</style>
+
+<center><h1> New Event </h1></center> 
+
+<!--Container for the event form-->
+<div class = "container">
+    <form action="/ConstructEvent" name="NewEvent" method="POST">
+        <!--Prompting for name of event-->
+        <label>What is the name of the Event?</label><br>
+        <input type="text" name="EventTitle" maxlength="50" required/><br>
+        <!--Prompting for day of event-->
+        <label>What day does this event occur?</label><br>
+        <input type="text" name="EventDay" maxlength="25" required/><br>
+        <!--Prompting for time of event-->
+        <label>What time does this event occur?</label><br>
+        <input type="text" name="EventTime" maxlength="15" required/><br>
+        <!--Prompting for location of the event-->
+        <label>Where is this event happening?</label><br>
+        <input type="text" name="EventLocation" maxlength="600" required/><br>
+        <!--Prompting for details of the event-->
+        <label>What is going on in this event?</label><br>
+        <input type="text" name="EventContent" maxlength="600" required/><br>
+        <!--Button to submit all inputted data of the event-->
+        <button type="submit">Create Event</button>   
+    </form>
+
+    <!--Cancellation button-->
+    <form action="/Community" method="GET">
+        <button type="submit">Cancel</button>
+    </form>
+</div>
+{{template "footer.html" .}}
