+++ Communities/app/views/CommunityEvents.html
@@ -1,16 +1,16 @@
 
 		<div class = "eventWindow">
-			<b> Brood War Tournament  </b><br/>
+			<b> Job Outcome </b><br/>
 			<i>Date : </i>
-			October 24, 2022
+			TBD
 			<br/>
 			<i>Time : </i>
-			9:00 A.M
+			TBD
 			<br/>
 			<i>Location : </i>
-			United Supermarkets Arena
+			Ft Worth
 			<br/>
 			<br/>
-			Join us for a good throwback to the 90's where RTS played a major role in our lives and the game first came out. Tournament will be a bracket style, so bring in your best builds! First place gets a Toshiba Liberetto and a $10,000 Grand Prize!
+			Waiting on whether or not I got the dream for doing software on Jets.
 		</div>
