+++ Communities/app/views/CommunityPosts.html
@@ -1,6 +1,6 @@
 
 		<div class = "postWindow">
-			<b> About those ASL Finals </b><br/>
-			Pros put up quite the show!
+			<b> What on earth happened? </b><br/>
+			Previously, this application ran ok with Latitude and Longitude being declared as floats in SQL. Now, for some reason, SQL gets pissed whenever floats are inserted. We had to change them to varchar as a result. Very Strange.
 		</div>
