+++ Communities/app/views/LatestPosts.html
@@ -1,0 +1,8 @@
+
+		<div class = "postWindow">
+			<b> What on earth happened? </b><br/>
+			Previously, this application ran ok with Latitude and Longitude being declared as floats in SQL. Now...<br/>
+			<br/>
+			<b>From Community Modular Bois</b>
+		</div>
+
