+++ Communities/none
@@ -455,3 +455,14 @@ INFO  2022/01/07 22:43:13  revel  revel.go:124: Paths                           
 INFO  2022/01/08 11:16:25  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
 INFO  2022/01/08 11:19:18  revel  revel.go:124: Paths                                   base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0"
 INFO  2022/01/08 12:46:41  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:25:47  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:31:25  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:34:35  revel  revel.go:124: Paths                                     revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"
+INFO  2022/01/08 13:36:20  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:37:10  revel  revel.go:124: Paths                                   views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app"
+INFO  2022/01/08 13:38:31  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:48:52  revel  revel.go:124: Paths                                    revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views" 
+INFO  2022/01/08 13:57:47  revel  revel.go:124: Paths                                   base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0"
+INFO  2022/01/08 14:01:39  revel  revel.go:124: Paths                                   base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0"
+INFO  2022/01/08 14:04:27  revel  revel.go:124: Paths                                   views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app"
+INFO  2022/01/08 14:05:41  revel  revel.go:124: Paths                                   revel="/home/eric/Recreational Projects/gocode/pkg/mod/github.com/revel/revel@v1.0.0" base="/home/eric/Recreational Projects/Capstone-Communities/Communities" app="/home/eric/Recreational Projects/Capstone-Communities/Communities/app" views="/home/eric/Recreational Projects/Capstone-Communities/Communities/app/views"
